refactor(downloads): consolidate iOS download helper into iosMain
The earlier metadata compilation failure that forced two identical actuals
in iosArm64Main and iosSimulatorArm64Main is gone once NSObject is imported
explicitly from platform.darwin (instead of relying on platform.Foundation.*).
:composeApp:compileIosMainKotlinMetadata now passes with the implementation
living entirely in iosMain, so the expect/actual indirection and the duplicate
target-specific files are no longer needed.
